Beginning implementation of Element Search
Element Search is one of the last major features for version 1.3. This creates the text box in the corner, although it currently doesn't do anything yet.

With this, I've also tried to make some other minor adjustments to the visuals to maybe help the program look better in high-DPI situations.