Add in Person Display contol (does nothing yet)
This is the beginnings of the last major feature for 1.3: the person display. This will be used to visually indicate where you have your items located and stored, and ideally also will be able to choose to equip and unequip them (I'll add on equiping UI to this and the item editor later on).

Right now, this is purely visual demonstration to help me determine how I want things to look; future commits will have me actually hooking up the side panes to actual data-driven items.

This is going to be an ongoing project for a bit... 1.3 will still be a bit of a ways away from now.