Created new controls to use with PersonDisplay
Thte PersonSlotItemList and SmalltemDisplay will be used with the PersonDisplay control. These will be the ways that I get the data into the control to actually properly display the items on the character sheet.

Still some progress to go, primarily in actually building the SmallItemDisplay and then hooking everything up in the PersonDisplay, but this is a major step forward.