Set up Person view for Equipment (not fully functional)
Set up the Person view to appear for Equipment/Gear, using the same visual appearance as the ability score view switcher.

Right now the Person view won't actually functionally do anything (still need to do a bit more on the SmallItemDisplay side), but it's at least there and looks nice.

Also set up the resizing methods for PersonView, so that when it's narrower, all the items end up on the right side.