Add Skill Points per Level, total Skill Points
These were added to aid with actually editing skills, as it was kind of hard to keep track of these numbers that otherwise are just not written down.

This marks a change in the PathfinderSheet object, so I'll need to also make sure I make this change with 1.2.4 for forwards compatibility.