Add support for enums as an IldType
This opens the door a bit wider for me to use enums as a variable type within these kinds of classes.

I plan on using this for a property type in the upcoming redesign of the SpellEditor, and it might have some good uses in future editors like SentinelsJson as well.