Add sort menu to ILD (doesn't work yet)
Added in a Sort menu and the beginnings of a function, but as of right now, it doesn't actually do anything.

This will be just used to rearrange all of the items currently in the display. I wanted to add it in since I thought it'd be helpful when dealing with a giant list of things (i.e. spells or items).